// src/app/impl/shared/model/third-party-rslt/third-party-rslt-h-x-obj.model.ts
export class ThirdPartyRsltHXObj {
    ThirdPartyRsltHId: number;
    SvcTypeCode: string;
    TrxTypeCode: string;
    TrxNo: string;
    CustName: string;
    IdType: string;
    IdNo: string;
    ReqDt: Date;
    ReqJson: string;
    ResultDt: Date;
    JobId: string;
    Consent: string;

    constructor() {
        this.Consent = null;
    }
}